static/requirements.txt
Django
Pillow
python-decouple
drf-yasg
djangorestframework
django-filter
psycopg2-binary
gunicorn
whitenoise
djangorestframework-simplejwt
social-auth-app-django
django-cors-headers
django-debug-toolbar
Celery
Redis
django-celery-results
channels
channels-redis